Interface do Profissional

+++ controle/ControladorProfissional.py
@@ -1,11 +1,11 @@
-from limite.TelaProfissional import TelaProfissional
+from limite_gui.TelaProfissionalGUI import TelaProfissionalGUI
 from entidade.profissional import Profissional
 from dao.ProfissionalDAO import ProfissionalDAO
 
 
 class ControladorProfissional:
     def __init__(self, controlador_principal):
-        self.__tela_profissional = TelaProfissional(self)
+        self.__tela_profissional = TelaProfissionalGUI(self)
         self.__profissional_dao = ProfissionalDAO()
         self.__controlador_principal = controlador_principal
 
@@ -16,31 +16,41 @@ class ControladorProfissional:
         return self.__profissional_dao.get(cpf)
 
     def selecionar_profissional_para_atendimento(self):
-        if len(self.__profissional_dao.get_all()) == 0:
-            self.__tela_profissional.mostrar_msg("Nenhum profissional cadastrado.")
+        profissionais = self.__profissional_dao.get_all()
+        
+        if len(profissionais) == 0:
+            self.__tela_profissional.mostrar_mensagem("Nenhum profissional cadastrado.")
             return None
-        self.listar_profissionais()
 
-        cpf = self.__tela_profissional.selecionar()
-        profissional = self.pegar_profissional_por_cpf(cpf)
+        cpf = self.__tela_profissional.tabela_profissionais(
+            profissionais,
+            selecionar=True
+        )
 
-        if profissional is None:
-            self.__tela_profissional.mostrar_msg("Profissional não encontrado.")
+        if cpf is None:
             return None
-        return profissional
+
+        if cpf < 0 or cpf >= len(profissionais):
+            self.__tela_profissional.mostrar_mensagem("Profissional não encontrado.")
+            return None
+        
+        return profissionais[cpf]
+    
 
     def selecionar_profissional_para_procedimento(self):
         return self.selecionar_profissional_para_atendimento()
 
     def incluir_profissional(self):
-        dados_profissional = self.__tela_profissional.pegar_dados()
-
+        dados_profissional = self.__tela_profissional.abrir_janela_cadastro()
+        if dados_profissional is None:
+            return
+        
         profissional_existente = self.pegar_profissional_por_cpf(
             dados_profissional["cpf"]
         )
 
         if profissional_existente is not None:
-            self.__tela_profissional.mostrar_msg("Profissional já cadastrado.")
+            self.__tela_profissional.mostrar_mensagem("Profissional já cadastrado.")
             return
 
         novo_profissional = Profissional(
@@ -52,29 +62,45 @@ class ControladorProfissional:
         )
 
         self.__profissional_dao.add(novo_profissional)
-        self.__tela_profissional.mostrar_msg("Profissional cadastrado com sucesso!")
+        self.__tela_profissional.mostrar_mensagem("Profissional cadastrado com sucesso!")
 
     def alterar_profissional(self) -> None:
         if len(self.__profissional_dao.get_all()) == 0:
-            self.__tela_profissional.mostrar_msg("Nenhum profissional cadastrado.")
+            self.__tela_profissional.mostrar_mensagem("Nenhum profissional cadastrado.")
             return
 
-        self.listar_profissionais()
+        cpf = self.__tela_profissional.tabela_profissionais(
+            self.__profissional_dao.get_all(),
+            selecionar=True
+        )
 
-        cpf = self.__tela_profissional.selecionar()
+        if cpf is None:
+            return
+        
         profissional = self.pegar_profissional_por_cpf(cpf)
 
         if profissional is None:
-            self.__tela_profissional.mostrar_msg("Profissional não encontrado.")
+            self.__tela_profissional.mostrar_mensagem("Profissional não encontrado.")
             return
+        
+        dados_antigos = {
+            "nome": profissional.nome,
+            "cpf": profissional.cpf,
+            "celular": profissional.celular,
+            "especialidade": profissional.especialidade,
+            "registro_profissional": profissional.registro_profissional
+        }
 
-        novos_dados = self.__tela_profissional.pegar_dados()
+        novos_dados = self.__tela_profissional.abrir_janela_cadastro(dados_antigos)
 
+        if novos_dados is None:
+            return
+        
         # Para não correr o risco de dois profissionais terem o mesmo CPF
         profissional_novo_cpf = self.pegar_profissional_por_cpf(novos_dados["cpf"])
 
         if profissional_novo_cpf is not None and profissional_novo_cpf != profissional:
-            self.__tela_profissional.mostrar_msg(
+            self.__tela_profissional.mostrar_mensagem(
                 "Já existe outro profissional com esse CPF."
             )
             return
@@ -88,61 +114,56 @@ class ControladorProfissional:
         profissional.registro_profissional = novos_dados["registro_profissional"]
 
         self.__profissional_dao.update(profissional)
-        self.__tela_profissional.mostrar_msg("Profissional alterado com sucesso.")
+        self.__tela_profissional.mostrar_mensagem("Profissional alterado com sucesso.")
 
     def excluir_profissional(self) -> None:
         if len(self.__profissional_dao.get_all()) == 0:
-            self.__tela_profissional.mostrar_msg("Nenhum profissional cadastrado.")
+            self.__tela_profissional.mostrar_mensagem("Nenhum profissional cadastrado.")
             return
 
-        self.listar_profissionais()
+        cpf = self.__tela_profissional.tabela_profissionais(
+            self.__profissional_dao.get_all(),
+            selecionar=True
+        )
 
-        cpf = self.__tela_profissional.selecionar()
+        if cpf is None:
+            return
+        
         profissional = self.pegar_profissional_por_cpf(cpf)
 
         if profissional is None:
-            self.__tela_profissional.mostrar_msg("Profissional não encontrado.")
+            self.__tela_profissional.mostrar_mensagem("Profissional não encontrado.")
             return
 
         self.__profissional_dao.remove(cpf)
-        self.__tela_profissional.mostrar_msg("Profissional removido com sucesso.")
+        self.__tela_profissional.mostrar_mensagem("Profissional removido com sucesso.")
 
     def listar_profissionais(self) -> None:
-        if len(self.__profissional_dao.get_all()) == 0:
-            self.__tela_profissional.mostrar_msg("Nenhum profissional cadastrado.")
+        profissionais = self.__profissional_dao.get_all()
+
+        if len(profissionais) == 0:
+            self.__tela_profissional.mostrar_mensagem("Nenhum profissional cadastrado.")
             return
-
-        for profissional in self.__profissional_dao.get_all():
-            dados_profissional = {
-                "nome": profissional.nome,
-                "cpf": profissional.cpf,
-                "celular": profissional.celular,
-                "especialidade": profissional.especialidade,
-                "registro_profissional": profissional.registro_profissional,
-            }
-
-            self.__tela_profissional.mostrar_profissional(dados_profissional)
+        
+        self.__tela_profissional.tabela_profissionais(profissionais)
 
     def abrir_tela(self) -> None:
-        lista_opcoes = {
-            1: self.incluir_profissional,
-            2: self.alterar_profissional,
-            3: self.excluir_profissional,
-            4: self.listar_profissionais,
-            0: self.voltar,
-        }
+        while True:
+            opcao = self.__tela_profissional.mostrar_menu()
+            if opcao == "Incluir Profissional":
+                self.incluir_profissional()
 
-        continua = True
+            elif opcao == "Alterar Profissional":
+                self.alterar_profissional()
 
-        while continua:
-            opcao_escolhida = self.__tela_profissional.mostrar_menu()
+            elif opcao == "Excluir Profissional":
+                self.excluir_profissional()
 
-            funcao_escolhida = lista_opcoes[opcao_escolhida]
+            elif opcao == "Listar Profissionais":
+                self.listar_profissionais()
 
-            if opcao_escolhida == 0:
-                continua = False
-
-            funcao_escolhida()
+            elif opcao == "Voltar":
+                break
 
     def voltar(self) -> None:
         return
